Enhance testing
